requirements.txt

# BarterHub - Platform Barter dengan Sistem Poin
# Requirements untuk Production dan Development

# Framework utama
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
WTForms==3.2.1

# Database dan ORM
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5

# Authentication dan Security
Werkzeug==3.0.1
PyJWT==2.8.0

# OAuth untuk login sosial
Flask-Dance==7.0.0
oauthlib==3.2.2

# Image processing untuk upload produk
Pillow==10.1.0

# Email validation
email-validator==2.1.0

# Production server
gunicorn==21.2.0

# Development dependencies
python-dotenv==1.0.0
python-decouple==3.8

# API Testing
requests==2.31.0

# Monitoring dan logging
sentry-sdk[flask]==1.38.0

# Rate limiting
Flask-Limiter==3.5.0

# CORS untuk API
Flask-CORS==4.0.0
requests
email-validator
flask
flask-login
flask-sqlalchemy
flask-wtf
gunicorn
pillow
psycopg2-binary
sqlalchemy
werkzeug
wtforms
email-validator
flask
flask-login
flask-sqlalchemy
flask-wtf
gunicorn
pillow
psycopg2-binary
werkzeug
wtforms
email-validator
flask
flask-login
flask-sqlalchemy
flask-wtf
gunicorn
pillow
psycopg2-binary
werkzeug
wtforms
